IF: Align with new interface and add CodedPhy support
The Phy now supports providing the coding_rate to the modem
to adjust the digital performance estimates for modulations
with variable conding rates.

Also add support for CodedPhy, by now by just offsetting
the 1Mbps with the approx. coding gain of Coded Phy
at 500 and 125kbps. The BER vs SNR shape and bit errors are not
going to be too correct, but it should be more than
sufficient for SW testing.

Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>